app/requirements.txt
altair==5.3.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
frozendict==2.4.4
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaleido==0.2.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multitasking==0.0.11
numpy==2.0.1
packaging==24.1
pandas==2.2.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
pretty-html-table==0.9.16
protobuf==5.27.3
psutil==6.0.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stockstats==0.6.2
streamlit==1.37.1
telegram==0.0.1
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
yfinance==0.2.41